pyproject.toml
[project]
name = "MangakAI"
version = "0.1.0"
description = "Transform your imagination into visual stories with the power of AI!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core AI and image processing
    "google-genai>=1.33.0",
    "pillow>=11.3.0",
    "reportlab>=4.4.3",
    
    # Web framework and API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "gunicorn>=21.2.0",
    "websockets>=12.0",
    
    # Database and ORM
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    
    # Async task processing
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # Cloud storage and utilities
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.1.1",
    
    # Monitoring and logging
    "prometheus-client>=0.19.0",
    "sentry-sdk[fastapi]>=1.38.0",
    
    # Development tools (optional)
    "gradio>=5.44.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
